// frontend/src/lib/api/client/generated/models/SortByDirection.ts
/*
 * Copyright 2025 sovity GmbH
 *
 * Licensed under the Apache License, Version 2.0 (the "License");
 * you may not use this file except in compliance with the License.
 * You may obtain a copy of the License at
 *
 * https://www.apache.org/licenses/LICENSE-2.0
 *
 * Unless required by applicable law or agreed to in writing, software
 * distributed under the License is distributed on an "AS IS" BASIS,
 * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 * See the License for the specific language governing permissions and
 * limitations under the License.
 *
 * SPDX-License-Identifier: Apache-2.0
 *
 * Contributors:
 *     sovity - init and continued development
 */
/* eslint-disable */
/**
 * sovity EDC API Wrapper
 * sovity\'s EDC API Wrapper contains a selection of APIs for multiple consumers, e.g. our EDC UI API, our generic Use Case API, our Commercial Edition APIs, etc. We bundled these APIs, so we can have an easier time generating our API Client Libraries.
 *
 * The version of the OpenAPI document: 0.0.0
 * Contact: contact@sovity.de
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * Direction of sort in list pages
 * @export
 */
export const SortByDirection = {
    Descending: 'DESCENDING',
    Ascending: 'ASCENDING'
} as const;
export type SortByDirection = typeof SortByDirection[keyof typeof SortByDirection];


export function instanceOfSortByDirection(value: any): boolean {
    for (const key in SortByDirection) {
        if (Object.prototype.hasOwnProperty.call(SortByDirection, key)) {
            if (SortByDirection[key as keyof typeof SortByDirection] === value) {
                return true;
            }
        }
    }
    return false;
}

export function SortByDirectionFromJSON(json: any): SortByDirection {
    return SortByDirectionFromJSONTyped(json, false);
}

export function SortByDirectionFromJSONTyped(json: any, ignoreDiscriminator: boolean): SortByDirection {
    return json as SortByDirection;
}

export function SortByDirectionToJSON(value?: SortByDirection | null): any {
    return value as any;
}

export function SortByDirectionToJSONTyped(value: any, ignoreDiscriminator: boolean): SortByDirection {
    return value as SortByDirection;
}